Move usb spdt and rudat implementations to private module
