+++ qcodes/instrument_drivers/Minicircuits/__init__.py
@@ -1,6 +1,6 @@
 from ._minicircuits_rc_sp4t import MiniCircuitsRCSP4T, MiniCircuitsRCSP4TChannel
 from ._minicircuits_rc_spdt import MiniCircuitsRCSPDT, MiniCircuitsRCSPDTChannel
-from .RUDAT_13G_90 import MiniCircuitsRudat13G90Usb
+from ._minicircuits_rudat_13g_90 import MiniCircuitsRudat13G90Usb
 from .USB_SPDT import MiniCircuitsSwitchChannelUsb, MiniCircuitsUsbSPDT
 
 __all__ = [
